Features/status2023 metadata

+++ src/egon/data/datasets/data_bundle/__init__.py
@@ -1,4 +1,4 @@
-"""The central module containing all code dealing with small scale inpu-data
+"""The central module containing all code dealing with small scale input-data
 """
 
 
@@ -13,7 +13,7 @@ from egon.data.datasets import Dataset
 
 def download():
     """
-    Download small scale imput data from Zenodo
+    Download small scale input data from Zenodo
     Parameters
     ----------
 
@@ -30,8 +30,10 @@ def download():
     )
     target_file = config.datasets()["data-bundle"]["targets"]["file"]
 
-    # Retrieve files
-    urlretrieve(url, target_file)
+    # check if file exists
+    if not Path(target_file).exists():
+        # Retrieve files
+        urlretrieve(url, target_file)
 
     with zipfile.ZipFile(target_file, "r") as zip_ref:
         zip_ref.extractall(".")
@@ -44,7 +46,7 @@ class DataBundle(Dataset):
         ]
         super().__init__(
             name="DataBundle",
-            version=f"{deposit_id}-0.0.1",
+            version=f"{deposit_id}-0.0.3",
             dependencies=dependencies,
             tasks=(download,),
         )
